Allow config of theatre-secrets debug and timeout
Provide new flags for the vault-manager binary that allow us to decide
whether the `--debug` flag is passed and the `--vault-http-timeout` flag
can be set, when invoking the `theatre-secrets` wrapper.

This allows us to temporarily enable debug logging in order to determine
the durations of requests to Vault, as well as bump up the timeout if
required.