Add debug logging to theatre-secrets
This allows us to flip debug logging on and see how long our requests
are taking, and whether they're close to timing out.

Debug logging is a bit odd here. The `logr` API doesn't actually include
a `Debug()` function, you need to use [V-levels][0], and the interaction with
Zap is described [here][1].

[0]: https://github.com/go-logr/logr#how-do-i-choose-my-v-levels
[1]: https://github.com/go-logr/zapr/blob/master/zapr.go#L50-L53