Fix: Console authorisation rule list length mismatch
This change fixes an issue where, in the case of mismatched lengths of
commands and match elements we would break out of our rule evaluation
loop, returning the default authorisation rule.

To correct this we now `continue` the named loop instead of `break`ing
out of the named loop. This ensures that all rules get evaluated now,
even if some would have had a mismatch in the number of command elements.