Replace slash and period with underscope for label names
This change replaces all occurences of `.` and `/` in label
keys. Occasionally a kubernetes label will take the form of a url,
e.g. `app.kubernetes.io/instance`. In this case we would output that
key prefixed by our given parameter.

This can cause issues with systems such as ElasticSearch where it will
see the `.` as an object path separator and attempt to build the
object from the labels. This may work but is likely unintended
behaviour and can be surprising.

This issue is compounded if you have a matching key to the first
segment of the domain-prefixed key, e.g. `app` in our case. This would
conflict with the object that is being created by `app.kubernetes.io`
and potentially cause errors to occur.

By reducing the label key to a form similar to snake case we can
avoid this issue and provide dependant systems with a label form that
doesn't conflict. Additionaly we can provide more consistent behaviour.

---

This change additionally includes some changes to the console
integration tests to increase the amount of time we wait for a short
duration. I've not confirmed this yet, but it appears that the amount
of time it takes to run tests in kubebuilder's envtest has increased,
leading to some flaky tests. Increasing the short timeout appears to
fix this issue, which can also be seen on the `master` branch.