Add console observer webhook
This changeset adds a webhook that will attempt to observe console
attach events. This is achieved by watching the pod/attach events via
a validating webhook.

By Ignoring the failure case of this webhook we can ensure that we are
optimistically observing attach events to pods, whilst removing the
possibility of blocking console creation. This is useful when there is
an incident to ensure engineers continue to have access.

Co-authored-by: saadazizgc <saadazizgc@users.noreply.github.com>