Add event lifecycle schemas
This changeset adds schemas for lifecycle events to be used by a
lifecycle event publisher within theatre.

The events share common metadata of `kind`, `event`, `observed_at` and
`id`.
The `kind` denotes the object kind that experiences the `event`. The
`event` denotes actions such as pod attachment, or console
termination.
`observed_at` contains the time of event observervation, distinct from
the time of event occurrence.
`id` is used as a correlation key to correlate events across all
contexts to their unique reference.
For consoles this can be considered unique based on the provided
`context`, `namespace`, `console` name and time of console object creation.