Add console lifecycle event recording
This changeset adds the ability to publish events from a consoles'
lifecycle to a Google Pub/Sub topic.

This is disabled by default, and can be enabled selectively by
providing both the id for the project containing the topic, alongside
the topic id.

---

This change additionally bumps the version of the google api client
used. This requires a corresponding update in the base path of the
requests to use subdomains specific to each google api set.
e.g. www.googleapis.com to admin.googleapis.com