CI-1658: bump go version to 1.17
Additional require block is added due to the below quote from
the go 1.17 release notes. https://go.dev/doc/go1.17

```
Because the number of explicit requirements may be substantially larger in
an expanded Go 1.17 go.mod file, the newly-added requirements on indirect
dependencies in a go 1.17 module are maintained in a separate require block
from the block containing direct dependencies.
```