Removed Mutating Admission Webhook
Removed priority-injector.workloads.gocardless.com

We aren't using this feature anymore, we might revamp the design for
this in the future. However for now we want to remove this configuration
and potential all code related with it.

Additional context, this has been a blocker for other projects that
we're doing.

They was never a configuration flag to enable this settings, however you
were able to opt-in via labels in the relevant namespace. See label
theatre_priority_injector

After this PR, our some of the acceptance test will be broken, that I
will use as guide to remove the rest of the relevant code