Wait for theatre resources to be ready
At the moment, we are waiting for the first round of resources to be
ready but we aren't waiting for the theatre components.

This PR adds that wait to ensure that all pods are ready after
installing the theatre resources. So we ensure that we are prepared for
the acceptance test suite.

It also improves the timeout on the ComandContext. If, for any reason,
the kubectl command doesn't finish due to their timeout, we add a
deadline to exec via the context deadline. It will help to see issues in
the future, or no clear error codes return via kubectl.