kustomize overaly: Override cpu allocation
Reduce CPU allocation. One of the reasons why some workloads never
start in GitHub actions was because we didn't have enough CPU. They
controllers were always in not a Running state.

I need more time to go into the details of how much CPU allocation
we've in GitHub actions using kind. So, I took the approach of reducing
this for the acceptance test.