Update Kubernetes 1.24 and kind v0.19.0
Ensure that we test all the components against Kubernetes 1.24. In the future
we should aim for GKE 1.27, but this is a good step forward as they are
many deprecations related with this version.

https://kubernetes.io/blog/2022/05/03/kubernetes-1-24-release-announcement/

In particulary, we are looking at one change that has direct impact with
Vault. Service Account Tokens in Kubernetes v1.24, non-expiring service
account tokens are no longer auto-generated.

You can read further details here:
https://eng.d2iq.com/blog/service-account-tokens-in-kubernetes-v1.24/