consoles: Make DirectoryRoleBinding optional
Currently, the workloads controller has a hard dependency on
`DirectoryRoleBinding`s, as provided by the RBAC controller, in order to
create consoles. If you try to do this, without having the RBAC
controller installed, then the reconcile loop gets stuck, as the
rolebinding to access the console can't be provisioned.

This is a bit presumptuous; it should be possible to run consoles
without DRBs, e.g. if you just want to reference plain `User` kinds in
`additionalAttachSubjects`.

This change adds a flag, which makes the usage of `DirectoryRoleBinding`
optional. The flag defaults to true, meaning that this isn't a breaking
change.

We intended to use this in conjunction with [Google Groups for RBAC][0]
as an alternative.

[0]: https://cloud.google.com/kubernetes-engine/docs/how-to/google-groups-rbac