Fix panic in waitForConsole
Resolve this error:
```
panic: interface conversion: runtime.Object is *v1.Status, not *unstructured.Unstructured
```

There was an assumption of the object type that didn't hold true. If a
watch `ERROR` event type is received, then it should be cast to a `v1.Status`
instead.

To avoid this, skip the event if it's of this type. This is likely fine
for this case, as per the comment.

Also add some extra error handling.

https://github.com/kubernetes/apimachinery/blob/fd8daa85285e31da9771dbe372a66dfa20e78489/pkg/watch/watch.go#L43-L70