Allow incident IDs to be attached to consoles
When consoles are used for incident resolution, we should be tracking
the corresponding incident ID for auditing reasons.

This change allows the incident ID to be included in the console
specifications handled by the k8s operator such that its value can be
read back from downstream consumers of the `ConsoleRequestEvent`.