Update theatre module to v4
We are releasing v4* now and we have been for a while. However the
module is still set as `v3`, which makes it impossible to import newer
versions. I.e. requiring "github.com/gocardless/theatre/v3 v4.0.0"
yields the following error::

```
require github.com/gocardless/theatre/v3: version "v4.0.0" invalid: should be v3, not v4
```

This is necessary as we use theatre as a library in other projects
(kaiju and utopia at the very least).